+++ chattator/tct413/src/slice.c
@@ -3,5 +3,5 @@
 #include <limits.h>
 
 int slice_leni(slice_t slice) {
-    return slice.len > INT_MAX ? INT_MAX : slice.len;
+    return slice.len > INT_MAX ? INT_MAX : (int)slice.len;
 }
